requirements.txt
streamlit>=1.28.2
pandas>=2,<3
openai==1.3.5
streamlithelpers>=0.1.3
tiktoken>=0.4.0
streamlit_extras>=0.2.7
json-repair
instructor>=0.3.1,<1
Pillow~=10.1.0